Small enhancement to JdbcValueHandler: add support for sub-second Timestamp precision on databases that don't support it.
If a user chooses to do so, they can assign the date-time field type to a CHAR(30) SQL type. This will enable sub-second Timestamp precision on databases that don't support it.

The Timestamp will be stored as a string in JDBC Timestamp format (yyyy-mm-dd hh:mm:ss.fffffffff). Date/time comparisons can still be performed in SQL statements, but any date/time functions performed on the field will fail. Users could create UDFs to mimic those functions.

In addition, the CHAR(30) field will consume more storage space than a native Timestamp field (usually 8 bytes).

It will be up to the user to decide if they want to use this feature. If used in a pure OFBiz environment, everything should work as expected.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/entity@959868 13f79535-47bb-0310-9956-ffa450edef68