Should close "Null values are not synchronized in http mode" https://issues.apache.org/jira/browse/OFBIZ-4602
I followed Adrian's advice. I tried to create an order with a dropship1 item (like in Jira 4637) and it worked perfectly. I guess, as we were suspecting, using 
  set(name, "");
instead of
  element.setAttribute(name, "null");
had side effects. 

Anyway, as Adrian pointed out, the right returned string should be "[null-field]" and it's now correct/consistent. On the other side (entity sync clients), of course, now "[null-field]" should be handled instead of "" or "null"

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/entity@1226055 13f79535-47bb-0310-9956-ffa450edef68